requirements-worker-gpu.txt
-r requirements-worker-common.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
onnxruntime-gpu
